infra: Make docker folder read only, remove unused tests, rename test-toolkit/ -> test/.
