+++ test/container/1.6.0/Dockerfile.mxnet.cpu
