+++ test/integration/__init__.py
@@ -16,19 +16,6 @@ import os
 
 RESOURCE_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'resources'))
 
-# these regions have some p2 instances, but not enough for automated testing
-NO_P2_REGIONS = [
-    'ca-central-1',
-    'eu-central-1',
-    'eu-west-2',
-    'us-west-1',
-    'eu-west-3',
-    'eu-north-1',
-    'sa-east-1',
-    'ap-east-1',
-    'me-south-1'
-]
-
 MODEL_SUCCESS_FILES = {
     'output': ['success'],
     'model': ['model-symbol.json', 'model-shapes.json', 'model-0000.params'],
