+++ test/integration/sagemaker/test_training.py
@@ -18,7 +18,7 @@ import pytest
 from sagemaker import utils
 from sagemaker.mxnet.estimator import MXNet
 
-from test.integration import RESOURCE_PATH
+from integration import RESOURCE_PATH
 from timeout import timeout
 
 DATA_PATH = os.path.join(RESOURCE_PATH, 'mnist')
@@ -26,7 +26,7 @@ SCRIPT_PATH = os.path.join(DATA_PATH, 'mnist.py')
 
 
 @pytest.mark.skip_test_in_region
-def test_training(sagemaker_session, ecr_image, instance_type, instance_count):
+def test_training(sagemaker_session, image_uri, instance_type, instance_count):
     hyperparameters = {'sagemaker_parameter_server_enabled': True} if instance_count > 1 else {}
     hyperparameters['epochs'] = 1
 
@@ -35,7 +35,7 @@ def test_training(sagemaker_session, ecr_image, instance_type, instance_count):
                train_instance_count=instance_count,
                train_instance_type=instance_type,
                sagemaker_session=sagemaker_session,
-               image_name=ecr_image,
+               image_name=image_uri,
                hyperparameters=hyperparameters)
 
     with timeout(minutes=15):
