--- test/resources/linear_regression/data/evaluation/eval_data.txt
@@ -1,96 +1,0 @@
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-7.000000000000000000e+00 2.000000000000000000e+00
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-6.000000000000000000e+00 1.000000000000000000e+01
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
-1.200000000000000000e+01 2.000000000000000000e+00
