--- test/resources/linear_regression/data/evaluation/eval_label.txt
@@ -1,96 +1,0 @@
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-1.100000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-2.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
-1.600000000000000000e+01
