+++ test/resources/mnist/mnist.py
@@ -93,19 +93,6 @@ def train(batch_size, epochs, learning_rate, num_gpus, training_channel, testing
                   num_epoch=epochs)
     if current_host == scheduler_host(hosts):
         save(model_dir, mlp_model)
-        assert_can_track_sagemaker_experiments()
-
-
-def assert_can_track_sagemaker_experiments():
-    in_sagemaker_training = 'TRAINING_JOB_ARN' in os.environ
-    in_python_three = sys.version_info[0] == 3
-
-    if in_sagemaker_training and in_python_three:
-        import smexperiments.tracker
-
-        with smexperiments.tracker.Tracker.load() as tracker:
-            tracker.log_parameter('param', 1)
-            tracker.log_metric('metric', 1.0)
 
 
 def save(model_dir, model):
