--- test/resources/onnx/mxnet_module/model-shapes.json
@@ -1,1 +1,0 @@
-[{"name": "data", "shape": [100, 1, 28, 28]}]
