--- test/resources/onnx/mxnet_module/model-symbol.json
@@ -1,111 +1,0 @@
-{
-  "nodes": [
-    {
-      "op": "null",
-      "name": "data",
-      "inputs": []
-    },
-    {
-      "op": "Flatten",
-      "name": "flatten0",
-      "inputs": [[0, 0, 0]]
-    },
-    {
-      "op": "null",
-      "name": "fullyconnected0_weight",
-      "attrs": {"num_hidden": "128"},
-      "inputs": []
-    },
-    {
-      "op": "null",
-      "name": "fullyconnected0_bias",
-      "attrs": {"num_hidden": "128"},
-      "inputs": []
-    },
-    {
-      "op": "FullyConnected",
-      "name": "fullyconnected0",
-      "attrs": {"num_hidden": "128"},
-      "inputs": [[1, 0, 0], [2, 0, 0], [3, 0, 0]]
-    },
-    {
-      "op": "Activation",
-      "name": "activation0",
-      "attrs": {"act_type": "relu"},
-      "inputs": [[4, 0, 0]]
-    },
-    {
-      "op": "null",
-      "name": "fullyconnected1_weight",
-      "attrs": {"num_hidden": "64"},
-      "inputs": []
-    },
-    {
-      "op": "null",
-      "name": "fullyconnected1_bias",
-      "attrs": {"num_hidden": "64"},
-      "inputs": []
-    },
-    {
-      "op": "FullyConnected",
-      "name": "fullyconnected1",
-      "attrs": {"num_hidden": "64"},
-      "inputs": [[5, 0, 0], [6, 0, 0], [7, 0, 0]]
-    },
-    {
-      "op": "Activation",
-      "name": "activation1",
-      "attrs": {"act_type": "relu"},
-      "inputs": [[8, 0, 0]]
-    },
-    {
-      "op": "null",
-      "name": "fullyconnected2_weight",
-      "attrs": {"num_hidden": "10"},
-      "inputs": []
-    },
-    {
-      "op": "null",
-      "name": "fullyconnected2_bias",
-      "attrs": {"num_hidden": "10"},
-      "inputs": []
-    },
-    {
-      "op": "FullyConnected",
-      "name": "fullyconnected2",
-      "attrs": {"num_hidden": "10"},
-      "inputs": [[9, 0, 0], [10, 0, 0], [11, 0, 0]]
-    },
-    {
-      "op": "null",
-      "name": "softmax_label",
-      "inputs": []
-    },
-    {
-      "op": "SoftmaxOutput",
-      "name": "softmax",
-      "inputs": [[12, 0, 0], [13, 0, 0]]
-    }
-  ],
-  "arg_nodes": [0, 2, 3, 6, 7, 10, 11, 13],
-  "node_row_ptr": [
-    0,
-    1,
-    2,
-    3,
-    4,
-    5,
-    6,
-    7,
-    8,
-    9,
-    10,
-    11,
-    12,
-    13,
-    14,
-    15
-  ],
-  "heads": [[14, 0, 0]],
-  "attrs": {"mxnet_version": ["int", 10300]}
-}
