+++ tox.ini
@@ -20,7 +20,6 @@ exclude =
     .tox
     tests/data/
     test/resources
-    test-toolkit/resources
     venv/
 max-complexity = 10
 ignore =
